moved interfaces to their own folder
