+++ src/examples/sum/interfaces/sum-agent-state.ts
@@ -1,4 +1,4 @@
-import { ISumEnvState } from "../../../src/examples/sum/sum-env-state";
+import { ISumEnvState } from "./sum-env-state";
 
 // tslint:disable-next-line:no-empty-interface
 export interface ISumAgentState extends ISumEnvState {}
