+++ src/examples/sum/interfaces/sum-data.ts
