+++ src/examples/sum/interfaces/sum-pop-options.ts
@@ -1,4 +1,4 @@
-import { IPopulationOptions } from "../../index";
+import { IPopulationOptions } from "../../../index";
 
 // tslint:disable-next-line:no-empty-interface
 export interface ISumPopOptions extends IPopulationOptions {
